Fix issue #8. Create new UI file from the autogenerated one.
The difference between the autogen file and this new one is that the
packages are imported, and no longer used with the classes.

ie.:
label = new javax.swing.JLabel();
  became
import javax.swing.JLabel
...
label = new JLabel();

In the future, when moving code from the autogen file to this one, we
should delete as well the packages from the code.