Get the user list when selecting a user.
Before, the user file list was requested when the user joined the sharing
service. This means that if the user adds a new file to its list, the
change is not visible unless he logs out and joins again.

Now, whenever a user selects another user, it sees the updated list. In
the future, the current simple read of the file list from the filesystem
should be implemented by a server request.